Add support for VM Pro by VO2Masters
Currently only reads data from device. Since the devices do not
remember setting between power cycles, it is importande that
the device if first configured using their app. When the
configuration is done, disconnect the device in the app so that
GC can connect to it.

Older firmwares do not support VCO2 measurement, and for those
devices it will be reported as zero. This also means that RER
will not be available since it's a function of VCO2 and VO2.