VMPro: Add VMPro specific window for configuration and detailed status
Added a separate window that pops up when a VMPro is detected, which can
be used to configure the device and monitor calibration progress and
show detailed status messages from the device. Most important info will
be forwarded to the small text field on the Train tab in the future, for
those with limited screen estate.

There's still some more settings to add, and a way to trigger a manual
calibration. I also want to add the possibility to save a detailed log
with the status messages from the device.